Renaming the connection branch to login
