+++ frontend/src/App.js
@@ -1,11 +1,16 @@
 import React from 'react';
-import Login from "./Login";
-
+import Login from "./login/login";
+import SecurianLogo from './login/securianLogo.png'
 
 function App() {
   return (
-      <div className="App">
-        <Login />
+      <div className="container">
+          <div className="login-logo-box">
+              <img src={SecurianLogo} alt="Securian Logo" />
+          </div>
+          <div className="login-box">
+              <Login />
+          </div>
       </div>
   );
 }
