--- frontend/src/IncorrectCredentialsBox.css
@@ -1,7 +1,0 @@
-.incorrect-credentials-box {
-    background-color: #ffcccc; /* Light red background color */
-    color: #cc0000; /* Dark red text color */
-    padding: 10px;
-    margin-top: 10px;
-    border-radius: 5px;
-}
