+++ frontend/src/login/incorrectCredentialsBox.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import './IncorrectCredentialsBox.css';
 
 function IncorrectCredentialsBox ({text}) {
     return <div className="incorrect-credentials-box">{text}</div>;
